Integration ebtween APP and Firebase complete
