+++ config/config.json
@@ -1,7 +1,7 @@
 {
   "development": {
     "username": "root",
-    "password": "mysql",
+    "password": "Developer01",
     "database": "roadmap_planner",
     "host": "127.0.0.1",
     "dialect": "mysql"
