+++ package.json
@@ -20,6 +20,8 @@
   "dependencies": {
     "express": "^4.17.1",
     "express-handlebars": "^5.0.0",
+    "firebase": "^7.16.1",
+    "jquery": "^3.5.1",
     "mysql2": "^2.1.0",
     "path": "^0.12.7",
     "sequelize": "^6.3.3"
