+++ views/login.handlebars
@@ -15,24 +15,11 @@
     </div>
 </section>
 
+
+
+<script src="https://www.gstatic.com/firebasejs/6.2.0/firebase-app.js"></script>
+<script src="https://www.gstatic.com/firebasejs/6.2.0/firebase-auth.js"></script>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js" type="text/javascript"></script>
-
-<script>
-
-$(document).ready(function(){
-    var emailLogin = $("#emailLogin");
-    var passwordLogin = $("#passwordLogin")
-    var submitLogin = $("#submitLogin");
-    console.log("Hello");
+<script src="js/login.js" type="text/javascript"></script>
 
 
-    submitLogin.on("click", ()=>{
-        var email = $("#emailLogin").val()
-        var password = $("#passwordLogin").val()
-
-        console.log("Email", email);
-        console.log("Password", password);
-    })
-
-})
-</script>
