+++ views/register.handlebars
@@ -13,11 +13,13 @@
                         <div class="row mt-4">
                             <div class="col">
                                 <input type="text" id="regEmail" class="form-control" placeholder="Email">
+                                <span class="helper-text" data-error="wrong" data-success="right"></span>
                             </div>
                         </div>
                         <div class="row mt-4">
                             <div class="col">
                                 <input type="password" id="regPassword" class="form-control mb-1" placeholder="Password">
+                                <span class="helper-text" data-error="wrong" data-success="right"></span>
                                 <p class="text-right mt-4 small ">
                                     <a href="/login">Already have an account?</a>
                                 </p>
@@ -35,26 +37,10 @@
     </div>
 </section>
 
+<script src="https://www.gstatic.com/firebasejs/6.2.0/firebase-app.js"></script>
+<script src="https://www.gstatic.com/firebasejs/6.2.0/firebase-auth.js"></script>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js" type="text/javascript"></script>
+<script src="js/login.js" type="text/javascript"></script>
 
-<script>
-
-$(document).ready(function(){
-
-    var submitBtn = $("#regSubmitBtn");
-
-    submitBtn.on("click", ( ) =>{
-    var name = $("#regName").val()
-    var email = $("#regEmail").val()
-    var password = $("#regPassword").val()
-
-    console.log("name", name);
-    console.log("email", email);
-    console.log("password", password)
-
-    })
-
-});
-</script>
